requirements.txt
flask
python-docx
fpdf
PyPDF2
